requirements.txt
streamlit==1.24.1
numpy==1.25.2
pillow==10.0.0
scikit-image==0.21.0
tensorflow==2.13.0
keras==2.13.1
opencv-python-headless==4.8.0.74